requirements.txt
Django
dj_database_url
djangorestframework
